Improve typing

+++ src/pycli/utils/math.py
@@ -1,17 +1,17 @@
 #!/usr/bin/env python3
 
 
-def add_numbers(a: int, b: int) -> int:
+def add_numbers(a: float, b: float) -> float:
     return a + b
 
 
-def sub_numbers(a: int, b: int) -> int:
+def sub_numbers(a: float, b: float) -> float:
     return a - b
 
 
-def mul_numbers(a: int, b: int) -> int:
+def mul_numbers(a: float, b: float) -> float:
     return a * b
 
 
-def div_numbers(a: int, b: int) -> float:
+def div_numbers(a: float, b: float) -> float:
     return a / b if b != 0 else float("inf")
